Split foundry lib and main
